Big change, can't get away as a lot of unit tests got affected by changing btree GetCurrentKey method to return Item struct instead of TK, to allow caller code to get the Item ID which is considered part of the Key, for B=tree non-unique. This is needed so we can implement SOP for Python being able to flow the item Key & the ID as generated internally in SOP, so then this can be used during "fetch".
SOP for Python GetKeys was also implemented & GetItems seen to return good values from the B-tree.
With the above changes, implementation of "GetValues" method can be made seamlessly bringing SOP for Python closer to code complete.

Also, it is identified that the current "context" implementation is sufficient. :)
If ever, a minor Python side "listener" equivalent of golang context will be created, or may be not. Depends on exploration findings. (looks like we're good)